split binding.hpp to multiple files
