--- engine/engine/nickel/script/internal/common.hpp
@@ -1,8 +1,0 @@
-﻿#pragma once
-#include "quickjs.h"
-
-namespace nickel::script {
-
-void LogJSException(JSContext*);
-
-}
