+++ engine/engine/nickel/script/script.hpp
@@ -10,6 +10,8 @@ public:
     ScriptManager();
     ~ScriptManager();
 
+    void Eval(std::string_view code);
+    void EvalBinary(std::span<uint8_t> code);
     QuickJSScript Load(const Path&);
     QuickJSScript Load(std::span<const char> content);
