+++ engine/engine/src/script/class.cpp
@@ -1,0 +1,5 @@
+#include "nickel/script/binding/class.hpp"
+
+namespace nickel::script {
+
+}
