+++ engine/engine/src/script/class_factory.cpp
@@ -1,0 +1,11 @@
+#include "nickel/script/binding/class_factory.hpp"
+
+namespace nickel::script {
+
+void QJSClassFactory::DoRegister() const {
+    for (auto& clazz : m_classes) {
+        clazz->EndClass();
+    }
+}
+
+}  // namespace nickel::script
