+++ engine/engine/src/script/script_impl.cpp
@@ -1,10 +1,31 @@
 #include "nickel/script/internal/script_impl.hpp"
 #include "nickel/context.hpp"
+#include "nickel/script/binding/common.hpp"
+#include "nickel/script/binding/runtime.hpp"
 
 namespace nickel::script {
 
 ScriptManagerImpl::ScriptManagerImpl() {
     bindingCpp();
+}
+
+void ScriptManagerImpl::Eval(std::string_view code) {
+    JS_Eval(QJSRuntime::GetInst().GetContext(), code.data(), code.size(),
+            nullptr, JS_EVAL_FLAG_STRICT | JS_EVAL_TYPE_GLOBAL);
+}
+
+void ScriptManagerImpl::EvalBinary(std::span<uint8_t> code) {
+    JSContext* ctx = QJSRuntime::GetInst().GetContext();
+    JSValue obj =
+        JS_ReadObject(ctx, code.data(), code.size(), JS_READ_OBJ_BYTECODE);
+
+    if (JS_IsException(obj)) LogJSException(ctx);
+
+    JSValue result = JS_EvalFunction(ctx, obj);
+    if (JS_IsException(result)) LogJSException(ctx);
+
+    JS_FreeValue(ctx, obj);
+    JS_FreeValue(ctx, result);
 }
 
 QuickJSScript ScriptManagerImpl::Load(const Path& filename) {
@@ -63,6 +84,14 @@ ScriptManager::ScriptManager()
     : m_impl{std::make_unique<ScriptManagerImpl>()} {}
 
 ScriptManager::~ScriptManager() {}
+
+void ScriptManager::Eval(std::string_view code) {
+    m_impl->Eval(code);
+}
+
+void ScriptManager::EvalBinary(std::span<uint8_t> code) {
+    m_impl->EvalBinary(code);
+}
 
 QuickJSScript ScriptManager::Load(const Path& filename) {
     return m_impl->Load(filename);
