+++ engine/engine/src/script/value_wrapper.cpp
@@ -1,0 +1,6 @@
+#include "nickel/script/binding/value_wrapper.hpp"
+
+namespace nickel::script {
+
+
+}
