+++ engine/tests/script/qjs/qjs_test.cpp
@@ -1,8 +1,16 @@
 #include "catch2/catch_test_macros.hpp"
 #include "nickel/context.hpp"
+#include "nickel/script/binding/module.hpp"
+#include "nickel/script/binding/runtime.hpp"
 #include "nickel/script/internal/script_impl.hpp"
 
 using namespace nickel;
+
+void ThrowWhenException(JSValue value) {
+    if (JS_IsException(value)) {
+        throw std::runtime_error("test failed");
+    }
+}
 
 struct Person {
     static int static_elem;
@@ -25,21 +33,71 @@ struct Person {
 
 int Person::static_elem = 255;
 
-TEST_CASE("test") {
-    script::ScriptManager mgr;
-    int int_elem = 1;
-    char char_elem = 2;
-    long long_elem = 3;
-    unsigned int uint_elem = 4;
-    unsigned char uchar_elem = 5;
-    unsigned long ulong_elem = 6;
-    float float_elem = 7;
-    double double_elem = 8;
-    bool bool_elem = true;
-    std::string_view str_view_elem = "string view";
-    const char* str_literal = "string literal";
-    std::string string_elem = "string";
+TEST_CASE("binding") {
+    SECTION("binding fundamental") {
+        int int_elem = 1;
+        char char_elem = 2;
+        long long_elem = 3;
+        unsigned int uint_elem = 4;
+        unsigned char uchar_elem = 5;
+        unsigned long ulong_elem = 6;
+        float float_elem = 7;
+        double double_elem = 8;
+        bool bool_elem = true;
+        std::string_view str_view_elem = "string view";
+        const char* str_literal = "string literal";
+        std::string string_elem = "string";
 
+        script::QJSRuntime runtime;
+        
+        auto& module =
+            runtime.GetContext().NewModule("test_module");
+
+        module.AddProperty("int_elem", int_elem)
+            .AddProperty("char_elem", char_elem)
+            .AddProperty("long_elem", long_elem)
+            .AddProperty("uint_elem", uint_elem)
+            .AddProperty("uchar_elem", uchar_elem)
+            .AddProperty("ulong_elem", ulong_elem)
+            .AddProperty("float_elem", float_elem)
+            .AddProperty("double_elem", double_elem)
+            .AddProperty("bool_elem", bool_elem)
+            .AddProperty("string_view", str_view_elem)
+            .AddProperty("str_literal", str_literal)
+            .AddProperty("string_elem", string_elem)
+            .EndModule();
+
+        script::QJSClassFactory::GetInst().DoRegister();
+
+        std::string_view code = R"(
+            function CheckExists(value) {
+                if (value == undefined || value == null) {
+                    throw new Error("value not exists");
+                }
+            }
+
+            CheckExists(test_module.int_elem)
+            CheckExists(test_module.uint_elem)
+            CheckExists(test_module.char_elem)
+            CheckExists(test_module.uchar_elem)
+            CheckExists(test_module.long_elem)
+            CheckExists(test_module.ulong_elem)
+            CheckExists(test_module.float_elem)
+            CheckExists(test_module.double_elem)
+            CheckExists(test_module.bool_elem)
+            CheckExists(test_module.string_view)
+            CheckExists(test_module.str_literal)
+            CheckExists(test_module.string_elem)
+        )";
+
+        JSValue value = JS_Eval(script::QJSRuntime::GetInst().GetContext(), code.data(),
+                code.size(), "test file",
+                JS_EVAL_TYPE_GLOBAL | JS_EVAL_FLAG_STRICT);
+
+        REQUIRE_FALSE(JS_IsException(value));
+    }
+
+    /*
     auto& module =
         script::QJSRuntime::GetInst().GetContext().NewModule("test_module");
     module.AddProperty("int_elem", int_elem)
@@ -90,4 +148,5 @@ TEST_CASE("test") {
     JS_FreeValue(ctx, global);
     JS_FreeValue(ctx, param);
     JS_FreeValue(ctx, fn);
+    */
 }
