+++ engine/tests/script/qjs/test.js
@@ -1,3 +1,4 @@
+
 Object.keys(test_module).forEach(key => {
     console.log(`Key: ${key}, Value: ${test_module[key]}`);
 });
